Turns out we had like 80% of the logic for a much cooler compare, with a minor Claude-assisted restructure
